refactor preview chat panel structure
