+++ src/features/preview/constants/chat.ts
@@ -1,0 +1,57 @@
+export const GOOGLE_FONTS = [
+  // Modern Sans
+  "Inter",
+  "Poppins",
+  "Roboto",
+  "Open Sans",
+  "Montserrat",
+  "Lato",
+  "Nunito",
+  "Raleway",
+  "Ubuntu",
+  "Work Sans",
+  "DM Sans",
+  "Manrope",
+  "Mulish",
+  "Rubik",
+  "Figtree",
+  "Outfit",
+  "Plus Jakarta Sans",
+
+  // Elegant / Luxury
+  "Playfair Display",
+  "Merriweather",
+  "Cormorant Garamond",
+  "Libre Baskerville",
+  "EB Garamond",
+  "Bodoni Moda",
+  "Prata",
+
+  // Bold / Creative
+  "Oswald",
+  "Bebas Neue",
+  "Anton",
+  "Archivo Black",
+  "Abril Fatface",
+  "Cinzel",
+
+  // Friendly / Startup
+  "Quicksand",
+  "Comfortaa",
+  "Josefin Sans",
+  "Cabin",
+  "Karla",
+  "Hind",
+  "Varela Round",
+
+  // Tech / Minimal
+  "Space Grotesk",
+  "Space Mono",
+  "IBM Plex Sans",
+  "JetBrains Mono",
+  "Source Sans 3",
+];
+
+export const ADD_NEW_PAGE = "add_new_page";
+
+export const MAX_REGEN_PROMPT_LENGTH = 1000;
